backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.56.0
django-cors-headers==4.2.0
django-crontab==0.7.1
django-filter==23.3
django-multiselectfield==0.1.13
djangorestframework==3.14.0
flake8
idna==3.4
oauthlib==3.2.2
pandas
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-barcode==0.15.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.4
unicode 
escpos